requirements/base.txt
# Bleeding edge Django
django==1.8.2

# Configuration
django-environ==0.3.0
django-secure==1.0.1
whitenoise==1.0.6


# Forms
django-braces==1.8.0
django-crispy-forms==1.4.0
django-floppyforms==1.4.1

# Models
django-model-utils==2.2

# Images
Pillow==2.8.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.20.0

# For the persistence stores
#psycopg2==2.6
MySQL-python

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.8.0

# Time zones support
pytz==2015.4

# Your custom requirements go here
python-cjson
simplejson
django-haystack
Whoosh
djangorestframework
django-reversion
jsonfield
selenium
lxml
cssselect
django-reversion-compare
diff-match-patch